refactor(artifact-provenance): extract finalization recovery
